added context, did intial migration, having issue displaying products

+++ Models/CustomerOrder.cs
@@ -8,15 +8,15 @@ namespace Shop_R_Us.Models
     public class CustomerOrder
     {
         public string Id { get; set; }
-        public decimal OrderTotal { get; set; }
-        public decimal SubOrderTotal { get; set; }
-        public decimal HighTaxTotal { get; set; }
-        public decimal LowTaxTotal { get; set; }
-        public decimal TaxTotal { get; set; }
+        public float OrderTotal { get; set; }
+        public float SubOrderTotal { get; set; }
+        public float HighTaxTotal { get; set; }
+        public float LowTaxTotal { get; set; }
+        public float TaxTotal { get; set; }
 
         public CustomerOrder () { }
 
-        public CustomerOrder(decimal orderTotal, decimal subOrderTotal, decimal highTaxTotal, decimal lowTaxTotal, decimal taxTotal)
+        public CustomerOrder(float orderTotal, float subOrderTotal, float highTaxTotal, float lowTaxTotal, float taxTotal)
         {
             OrderTotal = orderTotal;
             SubOrderTotal = subOrderTotal;
