+++ Models/Product.cs
@@ -10,11 +10,11 @@ namespace Shop_R_Us.Models
         public int Id { get; set; }
         public string ProductName { get; set; }
         public bool Fs { get; set; }
-        public decimal Price { get; set; }
+        public float Price { get; set; }
 
         public Product() { }
 
-        public Product(string productName, bool fs, decimal price)
+        public Product(string productName, bool fs, float price)
         {
             ProductName = productName ?? throw new ArgumentNullException(nameof(productName));
             Fs = fs;
